added catalog
so, here's the rub.
back when I first started working on this project, project gutenberg had a catalog available, which contained the books.  the catalog was their attempt at providing a way for automated users to search the site without triggering ratelimiting.  since then, gutenberg has removed the catalog from their site, and switched to some other method, which I'm not aware of.

this script works by grabbing the catalog from their site, but the logic for that grab had to be commented out, since they will ratelimit you if you attempt to grab the catalog.  I first found out about the change when my server was blacklisted.  not ratelimited, blacklisted.  I had to contact them to get unblacklisted.  of note, they aren't good at responding to messages, as in, they don't even respond at all.  but that's a side point.

I need to do some research to figure out what their new method of searching their site is, and how they want automated searchers to go about accessing their site.  to assume that all users are limited to looking through the web browser is a gross misunderstanding of the vast possiblity of users on the internet.  I do understand ratelimiting abusers or blatent scrapers of free ebooks, but some people, specifically me, use the site for more than just amassing a ton of plain text ebooks.

so, a rewrite of the script is in order at some point, to work with their new system.  for now, Ive added the catalog to the repo.  it's not updated with new ebooks, which is a problem, but with 40000 ebooks in the catalog, it's going to be a rare occurence to get a duplicate posted to the twitter site.