committed the script with commented out sections
as mentioned in the previous commit about the catalog, here is the commented out changes I've been locally running.
it's not ideal, but it works.

I will eventually rework the script itself to conform to gutenberg's standard and rule of access.  until then, it'll swirl around my mental project queue.