refactored commandline and added rc loading
testing mode is still set hard coded, because I didn't want to accidently post to twitter.  refactored that a little to make it a little more intuitive and informative.
created logic for alternative rc file, as well as added logic to verify (in simple terms) the rc config settings.
updated .gitignore