cleaned up catalog stuff
since the catalog isn't at gutenberg anymore, I removed the sub which attempted to download it, as well as the code which I had commented out.  No need for it to be in the script taking up space.

Ill find a new solution based on what gutenberg wants me to do, but for now, this works.