removed versioning
there isn't planned to be a release plan, with tiers, or any such thing for this script.
no need for versioning within the script.  all that can be maintained with commit history on git, and if I want, tags for releases.
I don't plan on reworking this majorly, outside of how it's getting the catalog.
maybe someday a major rework of the logic of the loop, but updates aren't dependent on backwards compatibility or anything like that.