adjusted options check to disallow -m and -t
admittedly, I found a bug.
if you run in manual mode, with twitter, it bombs out at the footer information.
not sure at first glance, but putting the need to fix this on my personal backlog.
tonight is just to get it live with the least amount of changes.
I dont want to stay up until midnight again working on stuff.