file handles were getting munged
so, the bug I mentioned earlier, wasn't just present when manual and twitter were specified, it happened whenever twitter was specified at all, which basically takes the majority of the purpose of this script away.

it took me a while to make the connection.  the clue was, when twitter was passed, I kept getting the following:
Use of uninitialized value in subroutine entry at quote.pl line 203, <$raw_fh>
but only when twitter was passed did my fh get lost.

I remembered from earlier I added code to read in the config using File::Slurper.
and for the book reading, I'm using open directly.
I could posit that File::Slurper isn't handling the file handle correctly, or that there is a variable naming collision, not really sure.
but that was the issue.

I removed File::Slurper and changed the config reading to use open, and it works again without issue.
I also removed the restriction for manual and twitter to run together.