added new download location
to remove the possibility of being ratelimited, I moved the download to their mirror instead of their main site.
there is still the matter of 404 not being accounted for, due to naming inconsistencies between the catalog and the mirror, but it's basically there, and much faster, just needs a little tweaking.

essentially, the naming convention with the mirror has a method to it's directories.  more information can be found in Issue #3.