removed UTF-8 read definition
Issue #8

Since the books can be in such a wide range of encodings, and don't always declare correctly, I removed the hard coded UTF8 definition.
Although this will allow non-translated UTF8 characters to be posted, this solves the issue of rogue encodings until I can integrate a system of properly detecting and converting the encoding.